server/requirements.txt
# Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
stripe==7.11.0
requests==2.31.0
